Canvas Bugs
After some signifigant trial and error, I have determined that the canvas is acting extremely weird. I will pick apart this bug using Widget for You ( my previous app that extensively used the canvas) in order to solve the problem. I believe I have a very good understanding of how to animate a bitmap but the canvas is just stubborn